app/requirements.txt
crispy-bootstrap5==0.7
Django==4.1.1
django-cors-headers==4.3.1
django-import-export==3.3.1
django-jazzmin==2.6.0
django-bootstrap5==23.1
django-crispy-forms==2.0
django-smart-selects==1.6.0
fontawesomefree==6.0
pandas==2.1.4
pytz==2023.3.post1
Pillow==9.2.0
django-tables2==2.4.1


# render config
python-dotenv==0.21.0
psycopg2-binary==2.9.5
dj-database-url==1.2.0
gunicorn==20.1.0
whitenoise==6.2.0
Brotli==1.0.9
